+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Improvements
 
+- New `#title` method for `Guard::Group` & `Guard::Plugin`. ([@rymai][])
 - Rename `Guard::Hook` to `Guard::Plugin::Hooker`. ([@rymai][])
 - New `Guard::Plugin::Base` module common to `Guard::Guard` (deprecated) & `Guard::Plugin`. ([@rymai][])
 - Deprecate `Guard::Guard` and replace it with `Guard::Plugin`. ([@rymai][])